+++ src/Collections/src/HashTable.php
@@ -12,7 +12,6 @@ declare(strict_types=1);
 
 namespace Aphiria\Collections;
 
-use ArrayIterator;
 use InvalidArgumentException;
 use OutOfBoundsException;
 use RuntimeException;
@@ -122,7 +121,7 @@ class HashTable implements IDictionary
      */
     public function getIterator(): Traversable
     {
-        return new ArrayIterator(\array_values($this->hashKeysToKvps));
+        return new KeyValuePairIterator(\array_values($this->hashKeysToKvps));
     }
 
     /**